+++ centreon/cypress/fixtures/apiTokens/creation/generatedTokenWithDefaultDuration.json
@@ -10,6 +10,6 @@
     },
     "token": "Yn5VaH9x3IJYjzKywO2XnhM5QITXQ4hqCvsXiqIwpZSkI\/jUobXBq14nDLAlFPvD",
     "creation_date": "2023-12-18T16:29:55+01:00",
-    "expiration_date": "2024-12-18T16:29:34+01:00",
+    "expiration_date": "2024-12-18T16:29:55+01:00",
     "is_revoked": false
 }
