+++ centreon/cypress/fixtures/apiTokens/listing/emptyList.json
@@ -1,0 +1,10 @@
+{
+    "result": [],
+    "meta": {
+      "page": 1,
+      "limit": 10,
+      "search": {},
+      "sort_by": { "token_name": "ASC"},
+      "total": 12
+    }
+  }
