+++ centreon/cypress/fixtures/apiTokens/listing/generatedToken.json
@@ -1,0 +1,25 @@
+{
+    "result": [
+      {
+        "name": "slack",
+        "user": {
+          "id": 4,
+          "name": "centreon-gorgone"
+        },
+        "creator": {
+          "id": 1,
+          "name": "admin admin"
+        },
+        "creation_date": "2023-12-18T16:29:55+01:00",
+        "expiration_date": "2024-12-18T16:29:55+01:00",
+        "is_revoked": false
+      }
+    ],
+    "meta": {
+      "page": 1,
+      "limit": 10,
+      "search": {},
+      "sort_by": { "token_name": "ASC"},
+      "total": 12
+    }
+  }
